:sparkles: upgrade dependency packages, rename path
