+++ frontend/.editorconfig
@@ -2,8 +2,8 @@
 root = true
 
 [*]
-indent_size = 2
 indent_style = space
+indent_size = 2
 end_of_line = lf
 charset = utf-8
 trim_trailing_whitespace = true
