+++ frontend/package.json
@@ -1,36 +1,31 @@
 {
   "name": "shizuru",
   "version": "1.0.0",
-  "description": "EKTotal frontend",
-  "author": "Nomuken <nomuken@spica.bz>",
+  "description": "shizuru is a EKTotal frontend",
+  "author": "nomuken",
   "private": true,
   "scripts": {
     "dev": "nuxt",
     "build": "nuxt build",
     "start": "nuxt start",
-    "generate": "nuxt generate",
-    "lint": "eslint --ext .js,.vue --ignore-path .gitignore .",
-    "precommit": "npm run lint"
+    "generate": "nuxt generate"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-free-regular": "^5.0.10",
-    "@fortawesome/fontawesome-free-solid": "^5.0.10",
-    "@nuxtjs/axios": "^5.3.1",
-    "@nuxtjs/toast": "^3.0.0",
-    "axios": "^0.18.0",
-    "bootstrap-vue": "^2.0.0-rc.8",
-    "nuxt": "^1.0.0",
-    "nuxt-fontawesome": "^0.1.3",
+    "@fortawesome/fontawesome-svg-core": "^1.2.17",
+    "@fortawesome/free-solid-svg-icons": "^5.8.1",
+    "@fortawesome/vue-fontawesome": "^0.1.6",
+    "@nuxtjs/axios": "^5.3.6",
+    "@nuxtjs/proxy": "^1.3.3",
+    "@nuxtjs/toast": "^3.1.0",
+    "bootstrap": "^4.1.3",
+    "bootstrap-vue": "^2.0.0-rc.11",
+    "cross-env": "^5.2.0",
+    "nuxt": "^2.4.0",
     "stylus": "^0.54.5",
     "stylus-loader": "^3.0.2",
-    "vue-clipboard2": "^0.1.0"
+    "vue-clipboard2": "^0.3.0"
   },
   "devDependencies": {
-    "@nuxtjs/proxy": "^1.2.4",
-    "babel-eslint": "^8.2.1",
-    "eslint": "^4.15.0",
-    "eslint-friendly-formatter": "^3.0.0",
-    "eslint-loader": "^1.7.1",
-    "eslint-plugin-vue": "^4.0.0"
+    "nodemon": "^1.18.9"
   }
 }
